requirements.txt
requests
pandas
numpy
transformers
torch
yfinance
newsapi-python
fastapi
uvicorn
streamlit
python-dotenv
pytest
black
flake8
sentencepiece
pydantic
plotly
mypy
python-dateutil